+++ packages/react/src/components/Button/RecentMessageButton/index.js
@@ -1,0 +1,1 @@
+export { default as RecentMessageButton } from './RecentMessageButton';
